# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/quality_procedures:
    get:
      tags:
        - QualityProcedures
      summary: List QualityProcedures
      description: Get a paginated list of QualityProcedures with optional filtering and sorting
      operationId: listQualityProcedures
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QualityProcedureList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - QualityProcedures
      summary: Create QualityProcedure
      description: Create a new QualityProcedure
      operationId: createQualityProcedure
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQualityProcedureRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QualityProcedure'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    QualityProcedure:
      type: object
      required:
        - id
        - company_id
        - procedure_name
        - is_active
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        procedure_name:
          type: string
          maxLength: 200
        parent_procedure_id:
          nullable: true
          type: string
          format: uuid
        description:
          nullable: true
          type: string
          maxLength: 4000
        is_active:
          type: boolean
        metadata:
          type: object
...
    CreateQualityProcedureRequest:
      type: object
      required:
        - company_id
        - procedure_name
        - is_active
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        procedure_name:
          type: string
          maxLength: 200
        parent_procedure_id:
          nullable: true
          type: string
          format: uuid
        description:
          nullable: true
          type: string
          maxLength: 4000
        is_active:
          type: boolean
        metadata:
          type: object
...
    QualityProcedureList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/QualityProcedure'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []